+++ Vish Mobile Backend/src/main/java/lk/ijse/vishmobilebackend/entity/TradePhonePhoto.java
@@ -9,10 +9,10 @@ import lombok.Data;
 public class TradePhonePhoto {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     @Column(name = "trade_phone_id")
-    private int phoneId;
+    private Long phoneId;
 
     @Column(name = "photo_url")
     private String photoUrl;
