+++ Vish Mobile Backend/src/main/java/lk/ijse/vishmobilebackend/repo/TradePhonePhotoRepo.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface TradePhonePhotoRepo extends JpaRepository<TradePhonePhoto, Integer> {
-    List<TradePhonePhoto> findByPhoneId(int phoneId);
+    List<TradePhonePhoto> findByPhoneId(Long phoneId);
 }
